Start refactor of Stats.js and add auto scroll to refactored layouts
